feat: univariate cauchy
